+++ packages/dartpollo/lib/dartpollo.dart
@@ -1,4 +1,14 @@
 export 'package:dartpollo_annotation/dartpollo_annotation.dart';
+export 'package:dio/dio.dart' show CancelToken, Dio;
+export 'package:gql_dio_link/gql_dio_link.dart'
+    show
+        DioLink,
+        DioLinkCancelTokenContextEntry,
+        DioLinkCanceledException,
+        DioLinkParserException,
+        DioLinkServerException,
+        DioLinkTimeoutException,
+        DioLinkUnkownException;
 
 export 'cache/cache_context.dart';
 export 'cache/cache_entry.dart';